requirements.txt
numpy
opencv-python 
matplotlib
python-dotenv
tqdm
setuptools